requirements.txt
torch
gymnasium
# gym==0.26.2  # legacy Gym for environments not migrated to Gymnasium
numpy
# pandas>=1.3 # use polars
polars
matplotlib
pyautogui
selenium
# flask # use fastapi
# flask-cors # use fastapi
fastapi[standard]
requests
ruff
webdriver-manager
Pillow
python-dotenv
pyperclip
stable-baselines3[extra]
opencv-python